src/requirements.txt
mysql-replication
kafka
six
